Move adapter dynamic metadata coverage to e2e
