+++ test/e2e/app-dir/adapter-dynamic-metadata/app/favicon.ico/route.ts
