+++ test/e2e/app-dir/adapter-dynamic-metadata/app/layout.tsx
