+++ test/e2e/app-dir/adapter-dynamic-metadata/app/robots.ts
