+++ test/e2e/app-dir/adapter-dynamic-metadata/my-adapter.mjs
