+++ test/e2e/app-dir/adapter-dynamic-metadata/next.config.js
@@ -1,0 +1,10 @@
+/**
+ * @type {import('next').NextConfig}
+ */
+const nextConfig = {}
+
+if (!process.env.NEXT_ADAPTER_PATH) {
+  nextConfig.adapterPath = require.resolve('./my-adapter.mjs')
+}
+
+module.exports = nextConfig
