--- test/production/app-dir/adapter-dynamic-metadata/next.config.js
@@ -1,8 +1,0 @@
-/**
- * @type {import('next').NextConfig}
- */
-const nextConfig = {
-  adapterPath: require.resolve('./my-adapter.mjs'),
-}
-
-module.exports = nextConfig
